move ActorModelController.tsx in domain actor
